+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * Add flexible clone method on virtual devices
 * Add possibility to disconnect all incoming/outgoing patches from a part or from the devices directly
 * Add possibility to easily remove decimals for the drag input (eaiser to deal with precision dragging)
+* Add experimental trigger schmitt neuron
+* Add experimental smooth min/max neuron
+
 
 ### Fixes